overcloud-data-gather.sh script
This script backs up the configuration files for all the OpenStack services, from all of the overcloud nodes. The following script creates a tar file of all the OpenStack configuration files for each overcloud node, and sends a copy of the tar file to the undercloud node’s /home/stack/backups directory.